Update fine tuning forms.

+++ Engine/Companions/ChatGPT/file.cs
@@ -2,7 +2,7 @@
 {
 	public class file : base_object
 	{
-		public int bytes { get; set; }
+		public long bytes { get; set; }
 		public string filename { get; set; }
 		public string purpose { get; set; }
 		public string status_details { get; set; }
